+++ Cpp02/ex03/includes/Point.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Point.hpp                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Helene <Helene@student.42.fr>              +#+  +:+       +#+        */
+/*   By: hlesny <hlesny@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 19:42:04 by hlesny            #+#    #+#             */
-/*   Updated: 2023/10/06 22:36:40 by Helene           ###   ########.fr       */
+/*   Updated: 2024/02/02 18:36:02 by hlesny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ class Point
 {
     public :
         Point();
-        Point(double const x, double const y);
+        Point(float const x, float const y);
         Point(Point const& to_copy);
         Point &operator=(Point const& u);
         bool operator==(Point const& u);
